fixed speech synth and added first draft one pager
I created a custom widget that is used to speak instead of attempting to use event handlers. Syntax is <<say string>> and it will speak the string. Sometimes the first word gets cut off but it only consistenly happens with the first passage that uses <<say>> I mitigate the problem by adding a trash word at the front that ends up not being said.